update exec and elf loader

+++ include/fat32.h
@@ -132,5 +132,5 @@ extern struct super_block * root_sb;
 
 void FAT32_init();
 struct dir_entry * path_walk(char * name,unsigned long flags);
-unsigned char* read_more_sector(unsigned int block, unsigned int count, unsigned int *size);
+unsigned char* read_more_sector(unsigned int block, unsigned int count, unsigned long *size);
 #endif
